Consolidate and relocate routes and models
All blueprints/routes and models now live in respectively-titled
files in the /api directory
