--- api/models/__init__.py
@@ -1,3 +1,0 @@
-from api.models.field import Field
-from api.models.task import Task
-from api.models.tool import Tool
