--- api/models/field.py
@@ -1,13 +1,0 @@
-from api import db
-
-class Field(db.Model):
-  __tablename__ = "field"
-  __table_args__ = {"mysql_charset": "utf8mb4"}
-
-  name = db.Column(db.String(80), primary_key=True, nullable=False)
-  description = db.Column(db.String(2047), nullable=False)
-  tasks = db.relationship("Task", backref="field", lazy="dynamic")
-  # input_type = db.Column(db.String(24), nullable=False)
-  input_options = db.Column(db.JSON(), nullable=True)
-  # constraints = db.Column(db.String, nullable=True)
-  pattern = db.Column(db.String(320), nullable=True)
